requirements.txt
<<<<<<< HEAD
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
=======
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
groq==0.13.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
>>>>>>> 1a0ec4e (subo primera inclusion de compu HP)
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
<<<<<<< HEAD
=======
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
typing_extensions==4.12.2
urllib3==2.3.0
>>>>>>> 1a0ec4e (subo primera inclusion de compu HP)
Werkzeug==3.1.3
